+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- 5 new distance/similarity
+	1. KuhnsIII
 ### Changed
 - PyPI badge in `README.md`
 - GitHub actions are limited to the `dev` and `master` branches